weakblocks: Implement "deltathin" transmission
This implements a more efficient transmission format in case
a block builds upon an already received weak block.

It checks the list of weak blocks and if the last weak block (or
2nd last, in case the last block is identical to the received block)
is, except for its Coinbase transaction, a subset of the block
in question, a special "deltathin" block will be produced that will
refer to the old block hash as the block to start assembling a new
block from. The transactions in the old block and new block must
match in order (except the Coinbase).

This transmission scheme only supports thinblocks, not _x_thinblocks,
due to the nature of the 64bit hash used for these and the fact that
mining zeroes out block hashes. If a deltathin block can be transmitted
this will take precedence over transmitting an xthin block.

On the sending side, in more detail:
------------------------------------
If the last weak block is detected as a subset of
the current block (with the partOfLastWeakBlock(..) function),
the vTxHashes array of the constructed CThinBlock is filled
with the hash of the "underlying" weak block first, followed by
the replacement coinbase transaction of the new block.

Note that this is currently an abuse of notation: vTxHashes
becomes "a weak block hash plus transaction hashes" instead of
only transaction hashes, as the name still indicates.

On the receiving side, in more detail:
--------------------------------------
If the first entry in the received CThinBlock's vTxHashes array
matches any of the currently tracked weakblocks, the transactions from
that weakblock (except the old coinbase) are taken for reconstruction,
but only after taking the second entry from vTxHashes for the new coinbase.