weakblocks: Re-request thinblocks that look like deltathin
Re-request a thinblock in case rebuilding it fails (merkle root
mismatch) but it looks like it might just be a deltathin block
from the hash value.

FIXME: Create a proper new block type "deltathin" to avoid this
"dual use of vTxHashes"-kludge.

Causing a rerequest of such a weak block as an attacker would mean
grinding a transaction (or any hash) to at least the weak difficulty
target. Which seems safe enough to guard against DOS, but the
general TODO of dealing with replay of old weakblocks might
apply here.