weakblocks: Move from latest seen to proper longest chain
This changes the weakblocks code from building on top of the latest
seen weakblock to build on top of the longest chain of weakblocks.

Other than regular strong blocks, delta- or weakblocks are refering
to each other only by implicitly including each others transactions.

For this reason, some machinery had to be build to automatically
figure out where a weak block fits into the weak blocks chain graph
and keep track of this graph in a proper way, even in the unlikely
but possible scenario that weakblocks arrive out-of-order.

A partial order relation '<' is defined on two weakblocks as such,
with wa, wb weak blocks:

wa < wb is false if wa == wb

wa < wb is false if wa contains the same number or more transactions
                 than wb

wa < wb is false if any transaction (except CB) in wa is not
                 in the same place in wb

Otherwise, wa < wb is true.

This relation is implemented in the weakExtends(..) function in the code
now.

The code internally keeps track of the the transitive reduction of the
 above '<' relation to a '<:' relation:

a <: b is true exactly when a < b and no other known block c exists
       for a!=c and b!= c and a<b<c.

This is called the miniextends (minimally extends) relation map in the code.
This represents the weak blocks chain graph (as per local node knowledge) and
the longest chain is derived from looking through this map.

The algorithm to insert still needs to be carefully checked for any problems
to maintain this order or similar. As a first test, a simple, randomized
'fuzzing' testing plus a scenario test will follow in one of the next
commits.

One potential problem of the current implementation might be the runtime
of the weakExtends(..) binary predicate. It currently has runtime O(n)
with n being the number of transactions in a block. By keeping a rolling
hash of the transactions (excl. CB) up to each point in the block to test
as well as a hash of all transactions in the blocks to be tested (those
that have been inserted already), it should be possible to reduce this
to ~O(1) amortized over all the repeated calls to weakExtends(..).
This has not been done yet, however.