weakblocks: Put underlying weak block hash into Coinbase
This adds logic to put the hash of the weak block a block builds upon
into the Coinbase. The hash is added as the second output using
a OP_RETURN construct.

On the receiving side, the outputs of the Coinbase transaction will be searched
for the first output that matches the pattern that sriptPubKey is "OP_RETURN <32-byte-data>"
and it is assumed that the data references the last weak block.