weakblocks: Update purge logic
This now marks all weak chain tips of height >=0 for removal during
the next step (sets the height to -1) and removes those chain tips
that already have a height of -1.

This causes removal of old weak chain tips to happen in two stages:
First step is marking, and second is removal.

This is important to still allow strong block transmission as
a delta block on top of the last weak block.