Use a single SHA256 hash in OP_CHECKDATASIG*
This changes OP_CHECKDATASIG and OP_CHECKDATASIGVERIFY to use just a
single execution of SHA256 to calculate the message hash instead of
hashing it twice using SHA256.

It also changes from serializing the hash using the Bitcoin serialization
library and the SER_HASH option to just hashing the raw bytes of the
corresponding value of the stack. This is to allow checking hashes of
signatures that do not follow this Bitcoin hashing convention. It should
also allow to emulate the old proposed behavior of OP_CHECKDATASIG* using
a combination of OP_CAT, the length and OP_SHA256 and manually building
serialized hashes according to this convention.

Note to reviewers: HMACs usually include the length of a message when
hashing, to prevent length extension attacks. Please also think about
whether similar things might be applicable here. I don't see how, but
I am not a crypto expert.

As OP_SHA256 would have similar issues, I don't think this is a worry. But
please spend some thoughts on this.

Test Plan: In an upcoming commit.