Import Bitcoin ABC's reworked signature encoding check function
Import this as IsValidSignatureEncodingWithoutSigHash(..). To be
used for CDS/-V.

This is the function that mainly resulted from this rework commit
in Bitcoin ABC:

49c7219314243398b9630ded78e97ebb0e3da776

Committer's notice: The author has been set to keep the attribution
of the author doing the rework. Also the signature has been changed
to work on vector<uint8_t>s again.

Co-authored-by: Awemany <awemany@protonmail.com>